requirements.txt
langchain
openai
tiktoken
faiss-cpu
python-dotenv
pypdf2
streamlit
huggingface_hub